+++ src/daemon/vhosts/hotel-dev.js
@@ -1,19 +1,9 @@
 const util = require('util')
 const express = require('express')
-const httpProxy = require('http-proxy')
-const conf = require('../../conf')
 
 // hotel.dev vhost
 module.exports = (servers) => {
   const app = express()
-  const proxy = httpProxy.createServer()
-  const target = `http://127.0.0.1:${conf.port}`
-
-  // Handle proxy error
-  proxy.on('error', (err, req, res) => {
-    util.log(err)
-    res.status(502).send(err)
-  })
 
   // Redirect http://hotel.dev/:id to http://:id.dev
   app.use('/:id', (req, res, next) => {
@@ -28,10 +18,10 @@ module.exports = (servers) => {
     next()
   })
 
-  // Delete host to avoid loop
-  app.use((req, res) => {
+  // Delete host to skip *.dev vhost handler
+  app.use((req, res, next) => {
     delete req.headers.host
-    proxy.web(req, res, { target })
+    next()
   })
 
   return app