+++ test/daemon/vhosts/dev.js
@@ -1,0 +1,34 @@
+/* global describe, after, it */
+const request = require('supertest')
+const app = require('../helper').createApp()
+
+describe('*.tld', () => {
+  after(done => app.shutdown(done))
+
+  describe('GET http://node.dev', () => {
+    it('should proxy request', done => {
+      request(app)
+        .get('/')
+        .set('Host', 'node.dev')
+        .expect(200, /Hello World/, done)
+    })
+  })
+
+  describe('GET http://unknown.dev', () => {
+    it('should return 404', done => {
+      request(app)
+        .get('/')
+        .set('Host', 'unknown.dev')
+        .expect(404, done)
+    })
+  })
+
+  describe('GET http://failing.dev', () => {
+    it('should return 502', done => {
+      request(app)
+        .get('/')
+        .set('Host', 'failing.dev')
+        .expect(502, done)
+    })
+  })
+})